search: Proof of Concept
Very simple and somewhat wonky proof-of-concept for search. This adds a
single "search.html" file. This files contains a number of javascript
functions that pull up the pre-generated search index and allow users to
search it dynamically using a webform.

+ index is generated as a pre-build step using the Gruntfile.js
+ dynamic searching is done using client-side javascript to search the
  index

The dynamic loading of results seems to work quite nicely. Currently the
links provided by the return function are broken. The fix here would
probably be to take a look at the Gruntfile and make sure the index gets
created correctly.

Signed-off-by: Adam McCartney <adam@mur.at>