Fix: Index generator
A couple of very minor changes to fix a problem with the parsing routine
used by the Gruntfile. The problem was the first index page was clogging
up the indexer because it had no yaml frontmatter. The indexer
explicitly looks to split the page based on its frontmatter. The last
commit had submitted a proof of concept with the bug of having broken
links to the search results.
This commit fixes that bug by making two small amendments:
+ adds frontmatter to the first _index.md so that it no longer jams up
  the indexer
+ use the correct prefix for the content directory, so the indexer
  creates correct hrefs

Signed-off-by: Adam McCartney <adam@mur.at>