Add production grade nginx
This commit adds the ability to serve the site via https locally. It
uses self-signed certificate approach. The nginx config was borrowed
from https://git.thelyoncompany.com/edwin/nginxconfig. And is quite
possibly overkill for what we need. A simple nginx http server would
have allowed us to serve the same thing. Anyway, this version shows how
to fully tweak out an nginx config. The nginx container is completely
stateless, so we can swap in and out the config at will, as well as the
site content.

Also included is a bash script to generate a set of self-signed
certificates using openssl, stolen shamelessly from:
https://devopscube.com/create-self-signed-certificates-openssl/

Signed-off-by: Adam McCartney <adam@mur.at>