Search: refactor with styling
This version of search builds on the proof of concept version by
integrating the feature more fully into the site. The resources
initially referenced to create the search feature were gleaned from:
https://gohugo.io/tools/search/.

Also the blog https://victoria.dev/ was extremely helpful.

This version is based more-or-less on a modifcation of
https://gist.github.com/sebz/efddfc8fdcb6b480f567 with some additions
and good ideas from the other sources mentioned above.

This commit also still features the XMLRequest api as we couldn't get
the fetch API to play nice for some reason. Probably we were trying to
make it too modular. There are two interdependent functions used to
generate the index and one of these is asynchronous (making them both
async). Our approach with fetch tried to get rid of the callback passing
style used in the XMLRequest version in this commit. Live and learn.
Some day we hope to understand how CPS actually works.

Signed-off-by: Adam McCartney <adam@mur.at>